ipv4/lwtunnel: add %LWTUNNEL_IP_FLAGS_BITMAP attribute
Add the ability to pass IP tunnel flags via Netlink bigint/bitmap
in IPv4 lwtunnel code between the userspace and the kernel. On
receiving, the neattribute is preferred, the old 16-bit one is being
read only if the new one is not present.
On sending, there's no other options other than sending both of
them, so that both old and new userspace programs could work
properly.

Signed-off-by: Alexander Lobakin <alexandr.lobakin@intel.com>