bitmap: try to optimize arr32 <-> bitmap on 64-bit LEs
Unlike bitmap_{from,to}_arr64(), when there can be no out-of-bounds
accesses (due to u64 always being no shorter than unsigned long),
it can't be guaranteed with arr32s due to that on 64-bit platforms:

bits     BITS_TO_U32 * sizeof(u32)    BITS_TO_LONGS * sizeof(long)
1-32     4                            8
33-64    8                            8
95-96    12                           16
97-128   16                           16

and so on.
That is why bitmap_{from,to}_arr32() are always defined there as
externs. But quite often @nbits is a compile-time constant, which
means we could suggest whether it can be inlined or not at
compile-time basing on the number of bits (above).

So, try to determine that at compile time and, in case of both
containers having the same size in bytes, resolve it to
bitmap_copy_clear_tail() on Little Endian. No changes here for
Big Endian or when the number of bits *really* is variable.

Signed-off-by: Alexander Lobakin <alexandr.lobakin@intel.com>