idpf: add function for hsplit configuring
The IDPF driver supports the header split feature and
that feature is always enabled by default.
However, for flexibility reasons and to simplify
the implementation of some features it would be useful
to have the support for switching the header split off.

Address that need by adding the user config parameter and
the function for disabling (or enabling) the header split
feature.

Signed-off-by: Michal Kubiak <michal.kubiak@intel.com>