idpf: reuse libie's definitions of parsed ptype structures
idpf's in-kernel parsed ptype structure is almost identical to the one
used in the previous Intel drivers, which means it can be converted to
use libie's definitions and even helpers. The only difference is that
it doesn't use a constant table, rather than one obtained from the
device.
Remove the driver counterpart and use libie's helpers for hashes and
checksums. This slightly optimizes skb fields processing due to faster
checks.

Signed-off-by: Alexander Lobakin <aleksander.lobakin@intel.com>